[Add]credit_card-edit&destroy

+++ app/controllers/users/addresses_controller.rb
@@ -21,10 +21,12 @@ class Users::AddressesController < ApplicationController
 
   def edit
     @address = Address.find(params[:id])
+    redirect_to root_path if current_user.id != @address.user.id
   end
 
   def update
     @address = Address.find(params[:id])
+    redirect_to root_path if current_user.id != @address.user.id
     if @address.update(address_params)
       flash[:success] = "編集が完了しました"
       redirect_to users_addresses_path
@@ -36,6 +38,7 @@ class Users::AddressesController < ApplicationController
 
   def destroy
     address = Address.find(params[:id])
+    redirect_to root_path if current_user.id != address.user.id
     address.destroy
     flash[:success] = "１件の配送先を削除しました"
     redirect_to users_addresses_path
